+++ app/api/prog_image/entities/image.rb
@@ -1,6 +1,7 @@
 module ProgImage
   module Entities
-    class Image < ImageKey
+    class Image < Grape::Entity
+      expose :key, documentation: { type: 'String', desc: 'Unique identifier for your image' }
       expose :public_url, documentation: { type: 'String', desc: 'URL image is located' }
     end
   end