+++ app/forms/prog_image/image_upload_form.rb
@@ -1,17 +1,12 @@
 module ProgImage
-  class ImageUploadForm < ::ProgImage::BaseForm
+  class ImageUploadForm < ImageBaseForm
     attr_reader :key
-
-    attribute :filename, String
-    attribute :type, String
-    attribute :head, String
-    attribute :tempfile
 
     validates :filename, format: { with: configatron.file.valid_filename_regex }
     validate :ensure_file_is_an_image
 
     def self.permitted_params
-      %i[filename type name tempfile head]
+      %i[filename type tempfile]
     end
 
     def persist
@@ -20,20 +15,8 @@ module ProgImage
 
     private
 
-    def upload_image
-      file_uploader.upload
-    end
-
     def ensure_file_is_an_image
       errors.add(:base) unless image_handler.image?
-    end
-
-    def image_handler
-      @image_handler ||= ImageHandler.new(tempfile)
-    end
-
-    def file_uploader
-      @file_uploader ||= FileUploader.new(attributes)
     end
   end
 end