+++ app/lib/prog_image/connectors/aws/s3_connector.rb
@@ -4,22 +4,35 @@ module ProgImage
   module Connectors
     module Aws
       class S3Connector < BaseConnector
-        def upload_file(file, key)
-          bucket.put_object file_upload_params(file, key)
+        include ProgImage::FileNaming
+
+        def fetch_file(key)
+          fetched_file = ::File.open("tmp/#{temporary_filename(key)}", 'wb') do |file|
+            client.get_object({ key: key, bucket: bucket_name }, target: file)
+          end
+          fetched_file.body
+        end
+
+        def fetch_file_url(key)
+          bucket.object(key).public_url
+        end
+
+        def file_exists?(key)
+          bucket.object(key).exists?
+        end
+
+        def upload_file(file_attributes, key)
+          bucket.put_object file_upload_params(file_attributes, key)
           true
         rescue ::Aws::S3::Errors, ::Timeout::Error
           false
         end
 
-        def fetch_file(key)
-          bucket.object(key)
-        end
-
         private
 
-        def file_upload_params(file, key)
+        def file_upload_params(file_attributes, key)
           {
-            body: file[:tempfile],
+            body: file_attributes[:tempfile],
             key: key,
             acl: 'public-read'
           }
@@ -27,6 +40,10 @@ module ProgImage
 
         def bucket
           @bucket ||= ::Aws::S3::Bucket.new region: region, credentials: credentials, name: bucket_name
+        end
+
+        def client
+          @client ||= ::Aws::S3::Client.new region: region, credentials: credentials
         end
 
         def bucket_name