+++ spec/lib/prog_image/connectors/aws/s3_connector_spec.rb
@@ -4,10 +4,29 @@ RSpec.describe ProgImage::Connectors::Aws::S3Connector do
   include_context 'with file for upload'
   include_context 'with aws s3 connector'
 
+  describe '#fetch_file' do
+    subject { connector.fetch_file key }
+
+    it { is_expected.to be_kind_of File }
+  end
+
+  describe '#fetch_file_url' do
+    subject { connector.fetch_file_url key }
+
+    it { is_expected.to be_kind_of String }
+    it { is_expected.to include key }
+  end
+
+  describe '#file_exists?' do
+    subject { connector.file_exists? key }
+
+    it { is_expected.to eq true }
+  end
+
   describe '#upload_file' do
     subject { connector.upload_file uploaded_image_file, key }
 
-    it { is_expected.to be_truthy }
+    it { is_expected.to be true }
 
     context 'when request returns error' do
       let(:expected_errors) { [::Aws::S3::Errors::ServiceError, Timeout::Error] }
@@ -18,14 +37,7 @@ RSpec.describe ProgImage::Connectors::Aws::S3Connector do
       end
 
       # flashing test
-      xit { is_expected.to be_falsy }
+      xit { is_expected.to be false }
     end
-  end
-
-  describe '#fetch_file' do
-    subject { connector.fetch_file key }
-
-    it { is_expected.to be_kind_of ::Aws::S3::Object }
-    its(:key) { is_expected.to eq key }
   end
 end