+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add Grape API Framework
 * Add image upload and retrieval endpoints
+* Extend retrieval endpoint to allow image extension
 
 #### Technical Improvements